Refactor dashboard-related config and routes
- In an attempt to start splitting up index.js into smaller, more
  descriptive files
